chore: create external hook
